% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordgen.R, R/zzz.R
\docType{package}
\name{chordgen}
\alias{chordgen}
\alias{chordgen}
\alias{chordgen-package}
\title{Generate random chords}
\usage{
chordgen(n, n_notes = 4, invert = TRUE, triad_types = 1:5,
  octave = seq(-1, 1, 1), transpose = seq(-5, 6, 1), add_lyric = TRUE)
}
\arguments{
\item{n}{number of chords}

\item{n_notes}{number of notes used for each chord. Between 3 and 7.}

\item{invert}{make inversions flag}

\item{triad_types}{1:major, 2:minor, 3:augmented, 4:diminished,
5:sus4, 6:m3d5}

\item{octave}{add octaves}

\item{transpose}{interval to make transpositions}

\item{add_lyric}{add text below chords?}
}
\description{
Generate random chords using many options

This is a random chord generator which uses music21 package
}
